# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for AI-Dia application",
        "title": "AI-Dia API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8005",
    "basePath": "/api/v1",
    "paths": {
...
        "/client/{client_id}/order-payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all order payments with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Payment"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pagination.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/client/{client_id}/order-payments/{payment_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get order payment details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Payment"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "payment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/order_payment.OrderPaymentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/client/{client_id}/order-payments/{payment_id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the status of an order payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Payment"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "payment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Payment Status Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order_payment.UpdatePaymentStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domains.PaymentStatus": {
            "type": "string",
            "enum": [
                "Unpaid",
                "Confirming_Payment",
                "Paid",
                "Refunded",
                "Voided"
            ],
            "x-enum-varnames": [
                "PaymentStatusUnpaid",
                "PaymentStatusConfirmingPayment",
                "PaymentStatusPaid",
                "PaymentStatusRefunded",
                "PaymentStatusVoided"
            ]
        },
...
        "helpers.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "order_payment.OrderPaymentResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expire_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "total_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "order_payment.UpdatePaymentStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "enum": [
                        "Unpaid",
                        "Confirming_Payment",
                        "Paid",
                        "Refunded",
                        "Voided"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domains.PaymentStatus"
                        }
                    ]
                }
            }
        },
        "pagination.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: \"Bearer {token}\" — paste the token from the login response (accessToken)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}